pyproject.toml
[project]
name = "themage"
version = "1.0.0"
description = "The Mage - A wise wizard assistant for NotesMage, powered by LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    # FastAPI and server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    # HTTP client for Express API calls
    "httpx>=0.27.0",
    # Environment and utilities
    "python-dotenv>=1.0.0",
    "langchain-groq>=1.1.1",
    "langchain-deepseek>=1.0.1",
]

[project.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
